requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
pymysql==1.0.2
pydantic<2.0
python-dotenv==0.19.0
email-validator==1.1.3